requirements.txt
pyyaml
psutil
readerwriterlock
Verilog_VCD
